refactor: step3 기능 우선 패키지 구성하기
tobe.kitchen 패키지에 도메인별로 패키지를 분리해봤습니다.
