+++ src/main/java/tobe/kitchenpos/menu/application/menugroup/MenuGroupService.java
@@ -1,0 +1,36 @@
+package tobe.kitchenpos.menu.application.menugroup;
+
+import tobe.kitchenpos.menu.domain.menugroup.MenuGroup;
+import tobe.kitchenpos.menu.domain.menugroup.MenuGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroup create(final MenuGroup request) {
+        final String name = request.getName();
+        if (Objects.isNull(name) || name.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        final MenuGroup menuGroup = new MenuGroup();
+        menuGroup.setId(UUID.randomUUID());
+        menuGroup.setName(name);
+        return menuGroupRepository.save(menuGroup);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MenuGroup> findAll() {
+        return menuGroupRepository.findAll();
+    }
+}
