--- src/main/java/tobe/kitchenpos/menu/ui/MenuRestController.java
@@ -1,60 +1,0 @@
-package tobe.kitchenpos.menu.ui;
-
-import tobe.kitchenpos.menu.application.MenuService;
-import tobe.kitchenpos.menu.domain.Menu;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import tobe.kitchenpos.menu.domain.MenuGroup;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-
-@RequestMapping("/api/menus")
-@RestController
-public class MenuRestController {
-    private final MenuService menuService;
-
-    public MenuRestController(final MenuService menuService) {
-        this.menuService = menuService;
-    }
-
-    @PostMapping
-    public ResponseEntity<Menu> create(@RequestBody final Menu request) {
-        final Menu response = menuService.create(request);
-        return ResponseEntity.created(URI.create("/api/menus/" + response.getId()))
-            .body(response);
-    }
-
-    @PutMapping("/{menuId}/price")
-    public ResponseEntity<Menu> changePrice(@PathVariable final UUID menuId, @RequestBody final Menu request) {
-        return ResponseEntity.ok(menuService.changePrice(menuId, request));
-    }
-
-    @PutMapping("/{menuId}/display")
-    public ResponseEntity<Menu> display(@PathVariable final UUID menuId) {
-        return ResponseEntity.ok(menuService.display(menuId));
-    }
-
-    @PutMapping("/{menuId}/hide")
-    public ResponseEntity<Menu> hide(@PathVariable final UUID menuId) {
-        return ResponseEntity.ok(menuService.hide(menuId));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<Menu>> findAll() {
-        return ResponseEntity.ok(menuService.findAll());
-    }
-
-    @PostMapping("/menuGroup")
-    public ResponseEntity<MenuGroup> menuGroupCreate(@RequestBody final MenuGroup request) {
-        final MenuGroup response = menuService.menuGroupCreate(request);
-        return ResponseEntity.created(URI.create("/api/menu-groups/" + response.getId()))
-                .body(response);
-    }
-
-    @GetMapping("/menuGroup")
-    public ResponseEntity<List<MenuGroup>> menuGroupFindAll() {
-        return ResponseEntity.ok(menuService.menuGroupFindAll());
-    }
-}
