+++ src/main/java/tobe/kitchenpos/menu/ui/menugroup/MenuGroupRestController.java
@@ -1,0 +1,31 @@
+package tobe.kitchenpos.menu.ui.menugroup;
+
+import tobe.kitchenpos.menu.application.menugroup.MenuGroupService;
+import tobe.kitchenpos.menu.domain.menugroup.MenuGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RequestMapping("/api/menu-groups")
+@RestController
+public class MenuGroupRestController {
+    private final MenuGroupService menuGroupService;
+
+    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+        this.menuGroupService = menuGroupService;
+    }
+
+    @PostMapping
+    public ResponseEntity<MenuGroup> create(@RequestBody final MenuGroup request) {
+        final MenuGroup response = menuGroupService.create(request);
+        return ResponseEntity.created(URI.create("/api/menu-groups/" + response.getId()))
+            .body(response);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<MenuGroup>> findAll() {
+        return ResponseEntity.ok(menuGroupService.findAll());
+    }
+}
