--- src/main/java/tobe/kitchenpos/order/domain/JpaOrderTableRepository.java
@@ -1,8 +1,0 @@
-package tobe.kitchenpos.order.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
-}
