--- src/main/java/tobe/kitchenpos/order/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package tobe.kitchenpos.order.domain;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
-}
