--- src/main/java/tobe/kitchenpos/order/domain/OrderType.java
@@ -1,5 +1,0 @@
-package tobe.kitchenpos.order.domain;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
