+++ src/main/java/tobe/kitchenpos/order/domain/common/JpaOrderTableRepository.java
@@ -1,0 +1,8 @@
+package tobe.kitchenpos.order.domain.common;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
