+++ src/main/java/tobe/kitchenpos/order/domain/common/OrderStatus.java
@@ -1,0 +1,5 @@
+package tobe.kitchenpos.order.domain.common;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
