+++ src/main/java/tobe/kitchenpos/order/domain/common/OrderTableRepository.java
@@ -1,0 +1,14 @@
+package tobe.kitchenpos.order.domain.common;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderTableRepository {
+    OrderTable save(OrderTable orderTable);
+
+    Optional<OrderTable> findById(UUID id);
+
+    List<OrderTable> findAll();
+}
+
