+++ src/main/java/tobe/kitchenpos/order/domain/common/OrderType.java
@@ -1,0 +1,5 @@
+package tobe.kitchenpos.order.domain.common;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
