+++ src/main/java/tobe/kitchenpos/order/domain/delivery/JpaDeliveryOrderRepository.java
@@ -1,0 +1,9 @@
+package tobe.kitchenpos.order.domain.delivery;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import tobe.kitchenpos.order.domain.common.Order;
+
+import java.util.UUID;
+
+public interface JpaDeliveryOrderRepository extends DeliveryOrderRepository, JpaRepository<Order, UUID> {
+}
