+++ src/main/java/tobe/kitchenpos/order/domain/eatin/EatInOrderRepository.java
@@ -1,0 +1,20 @@
+package tobe.kitchenpos.order.domain.eatin;
+
+import tobe.kitchenpos.order.domain.common.Order;
+import tobe.kitchenpos.order.domain.common.OrderStatus;
+import tobe.kitchenpos.order.domain.common.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface EatInOrderRepository {
+    Order save(Order order);
+
+    Optional<Order> findById(UUID id);
+
+    List<Order> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+}
+
