+++ src/main/java/tobe/kitchenpos/order/domain/takeout/JpaTakeOutOrderRepository.java
@@ -1,0 +1,9 @@
+package tobe.kitchenpos.order.domain.takeout;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import tobe.kitchenpos.order.domain.common.Order;
+
+import java.util.UUID;
+
+public interface JpaTakeOutOrderRepository extends TakeOutOrderRepository, JpaRepository<Order, UUID> {
+}
