+++ src/main/java/tobe/kitchenpos/order/ui/delivery/DeliveryOrderRestController.java
@@ -1,0 +1,57 @@
+package tobe.kitchenpos.order.ui.delivery;
+
+import tobe.kitchenpos.order.domain.common.Order;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import tobe.kitchenpos.order.application.delivery.DeliveryOrderService;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/orders/delivery")
+@RestController
+public class DeliveryOrderRestController {
+    private final DeliveryOrderService deliveryOrderService;
+
+    public DeliveryOrderRestController(final DeliveryOrderService deliveryOrderService) {
+        this.deliveryOrderService = deliveryOrderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Order> create(@RequestBody final Order request) {
+        final Order response = deliveryOrderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/delivery/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<Order> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(deliveryOrderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<Order> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(deliveryOrderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/start-delivery")
+    public ResponseEntity<Order> startDelivery(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(deliveryOrderService.startDelivery(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete-delivery")
+    public ResponseEntity<Order> completeDelivery(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(deliveryOrderService.completeDelivery(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<Order> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(deliveryOrderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Order>> findAll() {
+        return ResponseEntity.ok(deliveryOrderService.findAll());
+    }
+}
