+++ src/main/java/tobe/kitchenpos/order/ui/eatin/EatInOrderRestController.java
@@ -1,0 +1,47 @@
+package tobe.kitchenpos.order.ui.eatin;
+
+import tobe.kitchenpos.order.application.eatin.EatInOrderService;
+import tobe.kitchenpos.order.domain.common.Order;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+import java.util.UUID;
+
+@RequestMapping("/api/orders/eatin")
+@RestController
+public class EatInOrderRestController {
+    private final EatInOrderService eatInOrderService;
+
+    public EatInOrderRestController(final EatInOrderService eatInOrderService) {
+        this.eatInOrderService = eatInOrderService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Order> create(@RequestBody final Order request) {
+        final Order response = eatInOrderService.create(request);
+        return ResponseEntity.created(URI.create("/api/orders/eatin/" + response.getId()))
+            .body(response);
+    }
+
+    @PutMapping("/{orderId}/accept")
+    public ResponseEntity<Order> accept(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.accept(orderId));
+    }
+
+    @PutMapping("/{orderId}/serve")
+    public ResponseEntity<Order> serve(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.serve(orderId));
+    }
+
+    @PutMapping("/{orderId}/complete")
+    public ResponseEntity<Order> complete(@PathVariable final UUID orderId) {
+        return ResponseEntity.ok(eatInOrderService.complete(orderId));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Order>> findAll() {
+        return ResponseEntity.ok(eatInOrderService.findAll());
+    }
+}
